+++ wrappers/C/Testing/Suite_Core_Tests.cpp
@@ -131,17 +131,17 @@ SUITE(CORE_TESTS)
 
     TEST(GET_MICROSECONDS)
     {
-        double millis = 123;
-        double start = rr::getMicroSeconds();
+        int64_t millis = 123;
+        int64_t start = rr::getMicroSeconds();
         // sleep for milliseconds
         rr::sleep(millis);
-        double end = rr::getMicroSeconds();
-        double diff = end - start;
+        int64_t end = rr::getMicroSeconds();
+        int64_t diff = end - start;
 
         cout << "microseconds, start: " << start << ", end: " << end << ", diff: " << diff << endl;
 
         // timer varies from system to system, but should be semi-close
-        CHECK_CLOSE(diff, 1000. * millis, 2000.);
+        CHECK_CLOSE(diff, 1000 * millis, 2000);
 
         // make sure its increasing
         CHECK(diff > 0);
